fix: validate logged user for edit and delete routes

+++ src/controllers/ProjectController.js
@@ -15,6 +15,16 @@ module.exports = {
 
   async editProject(req, res) {
     const  { id, name }  = req.body
+
+    let user_id = req.user.id
+    
+    user_id = parseInt(user_id)
+
+    const projectValidation = await Project.findOne({ where: { id } })
+
+    if(projectValidation.user_id != user_id) {
+      return res.status(500).json("User invalid for this project !")
+    }
     
     const project = await Project.update({ name }, { where: { id }, returning: true })
 
@@ -33,6 +43,16 @@ module.exports = {
 
   async deleteProject(req, res) {
     const { id } = req.params
+    
+    let user_id = req.user.id
+    
+    user_id = parseInt(user_id)
+
+    const projectValidation = await Project.findOne({ where: { id } })
+
+    if(projectValidation.user_id != user_id) {
+      return res.status(500).json("User invalid for this project !")
+    }
 
     await Project.destroy({ where: { id } })
